requirements.txt
pybind11>=2.10.0
setuptools>=68.0.0
PySide6>=6.6.0
numpy>=1.24.0,<2.4
torch
torchvision
openai-whisper
faster-whisper
moviepy
Pillow
